backend/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.30.0
psycopg[binary]>=3.1.18

# Google Cloud
google-cloud-bigquery>=3.17.0
google-cloud-storage>=2.14.0

# Data processing
pandas>=2.2.0
numpy>=1.26.4
polars>=0.20.0

# API clients
httpx>=0.26.0
aiohttp>=3.10.0

# Validation & serialization
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
orjson>=3.9.15

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
